# Server/other/api/swagger.yaml
openapi: 3.0.1
info:
  title: SafeStreets API
  description: SafeStreets API for software engineering 2 project 2019/2020
  version: 1.0.0
servers:
- url: https://localhost:8080/v2
...
paths:
...
  /reports:
    get:
      tags:
      - dataAnalysis
      summary: Get a list of reports
      description: Returns list of reports, optionally filtered by type, city and
        date.
      operationId: reportsGET
      parameters:
      - name: from
        in: query
        description: From which date filter the content
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: to
        in: query
        description: To which date filter the content
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: type
        in: query
        description: Type filter to apply
        required: false
        style: form
        explode: true
        schema:
          type: string
      - name: city
        in: query
        description: City location filter to apply
        required: true
        style: form
        explode: true
        schema:
          type: integer
      responses:
        "200":
          description: List of reports
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/inline_response_200_2'
        "401":
          description: Not authenticated
          content: {}
      x-swagger-router-controller: DataAnalysis
...
components:
  schemas:
...
    Citizen:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
      example:
        firstname: Johnny
        lastname: De Gennaro
        email: asd@gmail.com
    Authority_info:
      type: object
      properties:
        firstname:
          type: string
        lastname:
          type: string
        email:
          type: string
      example:
        firstname: Bruno
        lastname: Bucciarati
        email: asd@pec.it
...
    Location:
      type: object
      properties:
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
      example:
        latitude: 10.1232
        longitude: 12.12323
    ViolationType:
      type: object
      properties:
        violation_type:
          type: string
          enum:
          - double_parking
          - invalid_handicap_parking
          - bike_lane_parking
          - red_zone_parking
          - parking_disk_violation
          - other
      example:
        violation_type: red_zone_parking
    ReportStatus:
      type: object
      properties:
        status:
          type: string
          enum:
          - pending
          - validated
          - invalidated
      example:
        status: validated
...
    Non_sensitive_data_report:
      example:
        timestamp: 2019-04-12T11:00:00Z
        city: Milan
        place: Via Golgi
      allOf:
      - type: object
        properties:
          timestamp:
            type: string
            format: date-time
          city:
            type: string
          place:
            type: string
      - $ref: '#/components/schemas/Location'
      - $ref: '#/components/schemas/ReportStatus'
      - $ref: '#/components/schemas/ViolationType'
    Non_sensitive_data_reports:
      type: array
      items:
        $ref: '#/components/schemas/Non_sensitive_data_report'
    Citizen_accessible_report:
      example:
        id: 0
        license plate: AA000AA
        photos:
        - path/image_1.jpg
        - path/image_2.jpg
      allOf:
      - $ref: '#/components/schemas/Non_sensitive_data_report'
      - type: object
        properties:
          id:
            type: integer
          license_plate:
            type: string
          photos:
            type: array
            items:
              type: string
              format: uri
...
    Report:
      example:
        submitter:
          firstname: Johnny
          lastname: De Gennaro
          email: asd@gmail.com
        supervisor:
          firstname: Bruno
          lastname: Bucciarati
          email: asd@pec.it
      allOf:
      - $ref: '#/components/schemas/Citizen_accessible_report'
      - type: object
        properties:
          submitter:
            $ref: '#/components/schemas/Citizen'
      - type: object
        properties:
          supervisor:
            $ref: '#/components/schemas/Authority_info'
    Reports:
      type: array
      items:
        $ref: '#/components/schemas/Report'
...
    inline_response_200_2:
      oneOf:
      - $ref: '#/components/schemas/Reports'
      - $ref: '#/components/schemas/Non_sensitive_data_reports'
...